id 및 image 변환 util화

+++ feature/pokemon/src/main/kotlin/com/pokemon/feature/pokemon/detail/DetailViewModel.kt
@@ -30,7 +30,7 @@ class DetailViewModel @Inject constructor(
                     state.copy(weight = it.weight, height = it.height, typeList = it.typeList)
                 }
                 it.moveList.forEach {
-                    getMoveDetail(moveId = it.toInt())
+                    getMoveDetail(moveId = it)
                 }
             }
         }
